Remove Unnecessary ViewModel and Display Selected Answer
